chore: Restructure project for building out components
