+++ src/components/atoms/Icon/Icons/Icons.ts
