+++ src/components/atoms/Icon/Icons/PauseCircle.tsx
