+++ src/components/atoms/Icon/Icons/StepBackward.tsx
