+++ src/components/atoms/Icon/Icons/StepForward.tsx
