+++ src/components/atoms/Icon/Icons/VolumeHigh.tsx
