+++ src/components/atoms/Icon/Icons/VolumeMedium.tsx
